refactor(cuda): drop redundant `_cuda` suffix from cuda submodules
Under `kornia_imgproc::cuda` the `cuda::` prefix already conveys the
backend, so the per-module suffix was redundant:

- cuda::color_cuda        → cuda::color
- cuda::resize_cuda       → cuda::resize
- cuda::warp_affine_cuda  → cuda::warp_affine

Pure module rename (files/dirs + all path references); item names
unchanged. `texture` was already unsuffixed.
