+++ crates/kornia-imgproc/examples/bench_cuda_color_conversions.rs
@@ -24,7 +24,7 @@ use cudarc::driver::{CudaContext, CudaSlice, CudaStream};
 use kornia_image::color_spaces::BayerPattern;
 use kornia_image::{Image, ImageSize};
 use kornia_imgproc::color::{self, ColormapType};
-use kornia_imgproc::cuda::color_cuda::{bayer, gray, hsv_hls, misc, swizzle, video, yuv};
+use kornia_imgproc::cuda::color::{bayer, gray, hsv_hls, misc, swizzle, video, yuv};
 
 const SIZES: &[(usize, usize)] = &[(640, 480), (1280, 720), (1920, 1080), (3840, 2160)];
 const N_BUFFS: usize = 8;
@@ -651,7 +651,7 @@ fn main() {
             &mut CudaSlice<f32>,
             usize,
         )
-            -> Result<(), kornia_imgproc::cuda::color_cuda::CudaColorError>;
+            -> Result<(), kornia_imgproc::cuda::color::CudaColorError>;
         let f32_cases: [(&'static str, CpuF32Fn, LaunchF32Fn); 2] = [
             (
                 "hsv_from_rgb_f32",
@@ -661,7 +661,7 @@ fn main() {
             (
                 "lab_from_rgb_f32",
                 color::lab_from_rgb,
-                kornia_imgproc::cuda::color_cuda::cie::launch_lab_from_rgb_f32,
+                kornia_imgproc::cuda::color::cie::launch_lab_from_rgb_f32,
             ),
         ];
         for (name, cpu_fn, launch_fn) in f32_cases {
