+++ crates/kornia-imgproc/src/cuda/color/cie.rs
@@ -511,7 +511,7 @@ mod tests {
     use kornia_image::{Image, ImageSize};
 
     use super::*;
-    use crate::cuda::color_cuda::test_utils::{default_stream, pattern_f32};
+    use crate::cuda::color::test_utils::{default_stream, pattern_f32};
 
     type CpuF64Fn = fn(&Image<f64, 3>, &mut Image<f64, 3>) -> Result<(), kornia_image::ImageError>;
     type Launch = fn(
@@ -546,7 +546,7 @@ mod tests {
     }
 
     fn assert_close(name: &str, gpu: &[f32], oracle: &[f32], tol: f32) {
-        let max_diff = crate::cuda::color_cuda::test_utils::max_abs_diff_f32(gpu, oracle);
+        let max_diff = crate::cuda::color::test_utils::max_abs_diff_f32(gpu, oracle);
         assert!(max_diff <= tol, "{name} max diff {max_diff} > {tol}");
     }
