+++ crates/kornia-imgproc/src/cuda/color/gray.rs
@@ -218,7 +218,7 @@ pub fn launch_rgb_from_gray_f64(
 #[cfg(all(test, feature = "cuda"))]
 mod tests {
     use super::*;
-    use crate::cuda::color_cuda::test_utils::{default_stream, pattern_f32, pattern_u8};
+    use crate::cuda::color::test_utils::{default_stream, pattern_f32, pattern_u8};
 
     // Odd size (tail coverage) + a realistic size.
     const SIZES: &[usize] = &[1, 37 * 23, 640 * 480];
@@ -255,7 +255,7 @@ mod tests {
         let gpu: Vec<f32> = stream.clone_dtoh(&d_dst).unwrap();
         stream.synchronize().unwrap();
 
-        let max_diff = crate::cuda::color_cuda::test_utils::max_abs_diff_f32(&gpu, &cpu);
+        let max_diff = crate::cuda::color::test_utils::max_abs_diff_f32(&gpu, &cpu);
         assert!(max_diff <= 1e-3, "f32 gray max diff {max_diff} > 1e-3");
     }
