+++ crates/kornia-imgproc/src/cuda/color/misc.rs
@@ -182,7 +182,7 @@ pub fn launch_apply_colormap_u8(
 #[cfg(all(test, feature = "cuda"))]
 mod tests {
     use super::*;
-    use crate::cuda::color_cuda::test_utils::{default_stream, pattern_u8};
+    use crate::cuda::color::test_utils::{default_stream, pattern_u8};
 
     #[test]
     fn colormap_bit_exact_vs_cpu() {
