+++ crates/kornia-imgproc/src/cuda/color/swizzle.rs
@@ -353,7 +353,7 @@ pub fn launch_rgb_from_rgba_u8(
 #[cfg(all(test, feature = "cuda"))]
 mod tests {
     use super::*;
-    use crate::cuda::color_cuda::test_utils::{default_stream, pattern_u8};
+    use crate::cuda::color::test_utils::{default_stream, pattern_u8};
 
     #[test]
     fn bgr_swap_u8_bit_exact_vs_cpu() {
